Add delayed indexing.
Manually re-index the document after it has a new revision.  For a new document
this will result in a double-index, but with just 1 revision it shouldn't be
too big of a performance hit.

For older documents with lots of revisions this might be a larger performance hit,
but I don't know any way to just do a delta index.  It's in the background anyways
so no real need to worry.